feat: encoder with more options [breaking changes]
